refactor: service, controller 통합
