+++ src/main/java/com/workat/api/mypage/service/MyPageService.java
@@ -1,12 +1,17 @@
 package com.workat.api.mypage.service;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.workat.api.mypage.dto.request.UserUpdateRequest;
 import com.workat.api.mypage.dto.response.MyProfileResponse;
+import com.workat.common.exception.BadRequestException;
 import com.workat.common.exception.ConflictException;
+import com.workat.common.exception.FileUploadException;
 import com.workat.common.exception.NotFoundException;
+import com.workat.common.util.FileUploadUtils;
 import com.workat.domain.user.entity.UserProfile;
 import com.workat.domain.user.repository.UserProfileRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +21,11 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @Service
 public class MyPageService {
+
+	@Value("${resources.upload-folder}")
+	private String resourcesLocation;
+	@Value("${resources.upload-uri}")
+	private String uploadUri;
 
 	private final UserProfileRepository userProfileRepository;
 
@@ -34,5 +44,28 @@ public class MyPageService {
 		UserProfile userProfile = userProfileRepository.findById(userId).orElseThrow(() -> new NotFoundException("워케이셔너가 존재하지 않습니다"));
 		userProfile.updateProfile(request.getNickname(), request.getPosition(), request.getWorkingYear(), request.getStory());
 		userProfileRepository.save(userProfile);
+	}
+
+	@Transactional
+	public String uploadProfileImage(Long userId, MultipartFile multipartFile) {
+		UserProfile userProfile = userProfileRepository.findById(userId).orElseThrow(() -> new NotFoundException("워케이셔너가 존재하지 않습니다"));
+
+		if (multipartFile.getOriginalFilename() == null || multipartFile.getOriginalFilename().isBlank()) {
+			throw new BadRequestException("빈 파일입니다.");
+		}
+		String savedFileName = null;
+		try {
+			savedFileName = FileUploadUtils.fileUpload(resourcesLocation, String.valueOf(userId), multipartFile.getBytes());
+			log.info("save actual path: " + savedFileName);
+
+			savedFileName = savedFileName.replace(resourcesLocation, uploadUri);
+			log.info("access image path: " + savedFileName);
+
+			userProfile.updateImage(savedFileName);
+			userProfileRepository.save(userProfile);
+		} catch (Exception e) {
+			throw new FileUploadException(e.getMessage());
+		}
+		return savedFileName;
 	}
 }
